+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Boolean keyvalues (yes/no) can now be set to `!$var`, to invert the value of the variable. 
 * Propcombine will now preserve prop fade distances, by calculating a new distance which encloses the original fade spheres.
 * Warn if propcombine or packing was disabled via command line.
+* Add option to `comp_adv_output` to have it expand target searches itself.
 
 ## Bugfixes
 * Fix a compile failure if prop ropes were placed in a group with no connections.